requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pinecone-client>=2.2.0
notion-client>=2.0.0
pandas>=1.5.0
plotly>=5.15.0
python-dotenv>=1.0.0
requests>=2.31.0